+++ R/LearnerTorchTabResNet.R
@@ -24,6 +24,9 @@
 #'   First dropout ratio.
 #' * `dropout2` :: `numeric(1)`\cr
 #'    Second dropout ratio.
+#' * `shape` :: `integer()` or `NULL`\cr
+#'   Shape of the input tensor. Only needs to be provided if the input is a lazy tensor with
+#'   unknown shape.
 #'
 #' @references
 #' `r format_bib("gorishniy2021revisiting")`
@@ -42,9 +45,12 @@ LearnerTorchTabResNet = R6Class("LearnerTorchTabResNet",
         classif = c("twoclass", "multiclass")
       )
 
+      check_shape = crate(function(x) check_shape(x, null_ok = TRUE, len = 2L))
+
       private$.param_set_base =  ps(
         n_blocks = p_int(0, tags = c("train", "required")),
-        d_block = p_int(1, tags = c("train", "required"))
+        d_block = p_int(1, tags = c("train", "required")),
+        shape = p_uty(tags = "train", custom_check = check_shape)
       )
       param_set = alist(private$.block$param_set, private$.param_set_base)
 
@@ -58,17 +64,36 @@ LearnerTorchTabResNet = R6Class("LearnerTorchTabResNet",
         callbacks = callbacks,
         loss = loss,
         man = "mlr3torch::mlr_learners.tab_resnet",
-        feature_types = c("numeric", "integer"),
+        feature_types = c("numeric", "integer", "lazy_tensor"),
       )
     }
   ),
   private = list(
     .block = NULL,
-    .dataset = function(task, param_vals) {
-      dataset_num(task, param_vals, argname = "num.input")
+    .ingress_tokens = function(task, param_vals) {
+      token = if (single_lazy_tensor(task)) {
+        shape = param_vals$shape %??% lazy_shape(task$head(1L)[[task$feature_names]])
+        if (is.null(shape)) {
+          stopf("Learner '%s' received task '%s' with lazy tensor feature '%s' with unknown shape. Please specify the learner's `shape` parameter.", self$id, task$id, task$feature_names) # nolint
+        } else if (is.null(param_vals$shape)) {
+          msg = check_shape(shape, len = 2L)
+          if (!isTRUE(msg)) {
+            stopf("Learner '%s' received task '%s' with lazy_tensor column of shape '%s', but the learner expects an input shape of length 2.", self$id, task$id, shape_to_str(shape))
+          }
+        }
+        ingress_ltnsr(shape = shape)
+      } else {
+        ingress_num(shape = c(NA, length(task$feature_names)))
+      }
+      list(input = token)
     },
     .network = function(task, param_vals) {
-      graph = po("torch_ingress_num", id = "num") %>>%
+      ingress = if (single_lazy_tensor(task)) {
+        po("torch_ingress_ltnsr", id = "num", shape = private$.ingress_tokens(task, param_vals)[[1L]]$shape)
+      } else {
+        po("torch_ingress_num", id = "num")
+      }
+      graph = ingress %>>%
         po("nn_linear", out_features = param_vals$d_block) %>>%
         po("nn_block", private$.block, n_blocks = param_vals$n_blocks) %>>%
         po("nn_head")